# sauron-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Seller:
    description: Information about sellers who offer parts and prebuilt firearms
    properties:
      affiliate_link_template:
        description: Template for creating affiliate links (NULL if not affiliate)
        example: https://www.brownells.com/?aff=gunguru_{product_id}
        type: string
      contact_info:
        description: Contact information for the seller
        example: '{"phone": "800-741-0015", "email": "support@brownells.com"}'
        type: string
      created_at:
        description: Creation timestamp
        type: string
      description:
        description: Description of the seller
        example: A trusted retailer of firearm parts and accessories.
        type: string
      id:
        description: Unique identifier for the seller
        example: 1
        type: integer
      is_affiliate:
        description: Whether the seller is an affiliate partner
        example: false
        type: boolean
      logo_url:
        description: URL to the seller's logo
        example: https://example.com/brownells_logo.png
        type: string
      name:
        description: Name of the seller
        example: Brownells
        type: string
      updated_at:
        description: Last update timestamp
        type: string
      website_url:
        description: Website URL of the seller
        example: https://www.brownells.com
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@sauron.io
    name: Sauron API Support
    url: https://sauron.io/support
  description: A comprehensive API for firearms database management including parts,
    models, categories, product listings, and compatibility.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sauron.io/terms/
  title: Sauron Backend API
  version: "2.0"
paths:
...
  /sellers:
    get:
      consumes:
      - application/json
      description: Get a list of all sellers in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Seller'
            type: array
      summary: Get all sellers
      tags:
      - Sellers
    post:
      consumes:
      - application/json
      description: Add a new seller to the database
      parameters:
      - description: Seller Info
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/models.Seller'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Seller'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new seller
      tags:
      - Sellers
  /sellers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific seller
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a seller
      tags:
      - Sellers
    get:
      consumes:
      - application/json
      description: Get details of a specific seller
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a seller by ID
      tags:
      - Sellers
    put:
      consumes:
      - application/json
      description: Update details of a specific seller
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Seller Info
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/models.Seller'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a seller
      tags:
      - Sellers
  /sellers/{id}/affiliate-status:
    patch:
      consumes:
      - application/json
      description: Update the affiliate status of a specific seller
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: integer
      - description: Affiliate Status Update Info
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update seller affiliate status
      tags:
      - Sellers
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"